splits iptables and nftables functionality
